muc stream-id are now guint16
20071024141823-7fe3f-660fda7ce7652149e0e371f564b4f9b9ac4a10d3.gz

+++ lib/gibber/gibber-bytestream-muc.h
@@ -59,7 +59,7 @@ GType gibber_bytestream_muc_get_type (void);
                               GibberBytestreamMucClass))
 
 void gibber_bytestream_muc_add_sender (GibberBytestreamMuc *bytestream,
-    const gchar *sender, guint8 stream_id);
+    const gchar *sender, guint16 stream_id);
 
 void gibber_bytestream_muc_remove_sender (GibberBytestreamMuc *bytestream,
     const gchar *sender);
